fix: split extension repair lifecycle code
